pyproject.toml
[tool.poetry]
name = "fast-training"
version = "0.1.0"
description = ""
authors = ["Note do Dan <danielmaiareis@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.32"
pydantic-settings = "2.3.4"
alembic = "^1.13.2"
pyproject-toml = "^0.0.10"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
tzdata = "^2024.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
factory-boy = "^3.3.1"

[tool.ruff]
line-length = 150
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W','PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev fast_training/app.py'
pre_test = 'task lint'
test = 'pytest -s --cov=fast_training -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
